Separate data loading to parent component and move form to children component
