+++ src/components/BlenderLogbook/components/FillingTile.tsx
@@ -14,7 +14,10 @@ import {
   PrimaryButton,
 } from '../../common/Button/Buttons';
 import { DropdownMenu, TextInput } from '../../common/Inputs';
-import { CommonTileProps, EMPTY_FILLING_EVENT_ROW } from '../BlenderLogbook';
+import {
+  CommonTileProps,
+  EMPTY_FILLING_EVENT_ROW,
+} from './BlenderFillingEventForm';
 
 type FillingEventRowProps = CommonTileProps & {
   index: number;
