+++ src/components/BlenderLogbook/components/PricingTile.tsx
@@ -1,7 +1,7 @@
 import { formatEurCentsToEur, mapGasToName } from '../../../lib/utils';
-import { CommonTileProps } from '../BlenderLogbook';
 import React from 'react';
 import { GasWithPricing } from '../../../lib/queries/gasQuery';
+import { CommonTileProps } from './BlenderFillingEventForm';
 
 type PricingTileProps = CommonTileProps & {
   gases: GasWithPricing[];
