+++ src/components/Logbook/components/SavingTile.tsx
@@ -1,8 +1,8 @@
 import { Field } from 'formik';
 import { ButtonType, PrimaryButton } from '../../common/Button/Buttons';
-import { LogbookCommonTileProps } from '../../BlenderLogbook/BlenderLogbook';
 import React from 'react';
 import styles from './SavingTile.module.scss';
+import { LogbookCommonTileProps } from '../LogBookBasicInfoTile';
 
 export const AirLogbookSavingTile: React.FC<LogbookCommonTileProps> = ({
   values,
